+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Bob versions changelog
 - Added integration tests for bobt (#648)
 
 #### Changed
+- Using interval logger in metric exporter to reduce error log density (#592)
+- Using standard Authorization header for basic auth (#616)
 - Change locks to sync where possible (#472)
 
 #### Fixed